+++ examples/resources/english/snapshot.json
