+++ examples/resources/portuguese/files/02885f77-6ad9-4d76-9a1e-294b6cbdca04.json
