+++ examples/resources/portuguese/files/5bd53e58-571b-4b2b-b737-730f8435f7cc.json
