+++ examples/resources/portuguese/files/7db7112c-927c-45aa-8665-0b9b158d89ea.json
