+++ examples/resources/portuguese/files/980b249c-e8aa-4eb6-8049-bfc8b140e44b.json
