+++ examples/resources/portuguese/files/9c2c525d-4f9c-4517-9b8a-11a368811874.json
