+++ examples/resources/portuguese/files/d086ea04-b507-4386-aaf8-ca1ce344732a.json
