+++ examples/resources/portuguese/snapshot.json
