+++ examples/resources/spanish/snapshot.json
